Fix download card getting stuck.
In some scenarios, when a download fails (or hangs), the notification dissapears from the status bar although it's still in the download manager and after some time it gives a download complete but it really failed.
This change gives the user the chance to cancell the download in this case.

Also, i've fixed the carriage returns (lines added are 89-93)

Change-Id: I5f0203baa6dbc16b2592c74052896bbed02f0de1