OTA: Clean up & fixes
OTA: Make sure clean-up doesn't mind comments

When running formatting cleanup on the project, comments are
automatically compressed which means that the source is left with
a chunk that is hard to understand. This commit leaves the same
contents of the comments, but simply adds linebreaks and paragraphs
at places where they were already intended.

Change-Id: I82107dd30a3bd7db2866194fc39a3c39fcff7c25
OTA: Apply additional fancy code styling

On the second run, Eclipse decided to provide additional code
styling tips by replacing 4 spaces of padding with 8 spaces for
array initialization.

Change-Id: I569d68666acc39e7d3c841a652d84108a4acfaf4
OTA: Fix a widespread NPE during state saving

The package information for the latest updates could be left unset
(null) in case the update process fails which results in a NPE being
thrown when it the value is interacted with later on. This commit
implements a simple fix to ensure that in the worst case scenario,
when the "last updates" value is requested to be set as null, it is
set as an empty array which allows it to be freely used without the
risk of a NPE being thrown.

Fixes AOSPA-642.

Change-Id: Ifb06583562e3a4891170d17622118fe1f6877dc5
OTA: Clean up line endings

The line endings in the source code should be consistent throughout
the app. This commit swaps out the Windows/DOS style (CR+LF) endings
to instead use the UNIX style (LF) endings.

Change-Id: I55dcd9d5afce7b3330c2d493f5dc4193b950c4b2
OTA: Clean up code formatting throughout the app

Eclipse code formatting tool proposes a bunch of changes to clean
up the code styling. This is to provide a codebase that follows
consistent styling rules and is good looking, which in turn allows
it to be worked on easier.

Change-Id: I484eb1927adfecba9d05f32e9b8a31e4a82ed722